+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Anticipated release: October 26th, 2020
 
 #### 🚀 New features
 
+- Update Program Activities headers, labels, and help text ([#2528])
 - Rename State Cost Category page and update Nav Bar to better assist users in navigation ([#2432])
+- Rename Objectives and Key Results to Outcomes and Metrics ([#2584])
 
 #### 🐛 Bugs fixed
 
@@ -14,4 +16,6 @@ Anticipated release: October 26th, 2020
 
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
+[#2528]: https://github.com/CMSgov/eAPD/issues/2528
 [#2432]: https://github.com/CMSgov/eAPD/issues/2432
+[#2584]: https://github.com/CMSgov/eAPD/issues/2584